refactor: modify the TicketOffice class
