+++ backend/station_manager/app/src/main/java/com/StationManager/app/domain/train_station/Direction.java
@@ -1,4 +1,4 @@
-package com.StationManager.app.domain.trainstation;
+package com.StationManager.app.domain.train_station;
 
 public enum Direction {
     Up,
