+++ backend/station_manager/app/src/test/java/com/StationManager/app/domain/train_station/TicketOfficeTest.java
@@ -1,4 +1,4 @@
-package com.StationManager.app.domain.trainstation;
+package com.StationManager.app.domain.train_station;
 
 import static org.junit.jupiter.api.Assertions.*;
